refactor protocol module code more clearly
